replace TreeLoader by renaming EventTreeLoader to TreeLoader
