+++ lib/loaders/TreeLoader/closeChild.js
