+++ lib/loaders/TreeLoader/closeChildren.js
@@ -18,21 +18,6 @@ module.exports = function (parentId) {
   const kidIds = this.tracker(parentId, parentSpace)
 
   kidIds.forEach(kid => {
-    // Remove possible placeholders.
-    const placeholder = this.placeholders[kid]
-    if (placeholder) {
-      delete this.placeholders[kid]
-      placeholder.remove()
-    }
-
-    // Remove possible spaces.
-    const kidSpace = this.spaces[kid]
-    if (kidSpace) {
-      // Close its children. Note recursion.
-      this.closeChildren(kid)
-      // Then remove
-      delete this.spaces[kid]
-      kidSpace.remove()
-    }
+    this.closeChild(parentId, kid)
   })
 }
