+++ lib/loaders/TreeLoader/closeNeighbors.js
@@ -51,13 +51,29 @@ module.exports = function (rootIds, maxDepth) {
     }
   })
 
+  const closed = []
+
   findings.forEach(finding => {
     if (finding.depth > maxDepth) {
+      // Finding is outside the max depth threshold.
       const space = this.spaces[finding.id]
       if (space) {
+        delete this.demand[finding.id]
         delete this.spaces[finding.id]
         space.remove()
+        closed.push(finding.id)
       }
+    } else {
+      // Finding is inside the max depth threshold.
+      // Limit the demand for possible ongoing loading process.
+      const currentDemand = this.demand[finding.id] || 0
+      const maxDemand = maxDepth + 1 - finding.depth
+      this.demand[finding.id] = Math.min(currentDemand, maxDemand)
     }
   })
+
+  // Enable host app to hook to removals
+  for (let i = 0; i < closed.length; i += 1) {
+    this.emit('close', { id: closed[i] })
+  }
 }
