+++ lib/loaders/TreeLoader/closeParent.js
@@ -8,6 +8,7 @@ module.exports = function (childId) {
   //     a string
   //
 
+  // The child may be already closed.
   const childSpace = this.spaces[childId]
   if (!childSpace) {
     return
@@ -16,20 +17,29 @@ module.exports = function (childId) {
   // Find parent.
   const parentId = this.backtracker(childId, childSpace)
 
-  // Remove possible placeholders.
-  const placeholder = this.placeholders[parentId]
-  if (placeholder) {
-    delete this.placeholders[parentId]
-    placeholder.remove()
+  // Parent may be already closed.
+  const parentSpace = this.spaces[parentId]
+  if (!parentSpace) {
+    return
   }
 
-  // Remove possible spaces.
-  const parentSpace = this.spaces[parentId]
-  if (parentSpace) {
-    // Recursion. Close its parents.
-    this.closeParent(parentId)
-    // Then remove
-    delete this.spaces[parentId]
-    parentSpace.remove()
-  }
+  // Close siblings
+  const childIds = this.tracker(parentId, parentSpace)
+
+  childIds.forEach((kid) => {
+    if (kid !== childId) {
+      this.closeChild(kid)
+    }
+  })
+
+  // Close the parent's parent. Note the recursion.
+  this.closeParent(parentId)
+
+  // Then remove
+  delete this.demand[parentId]
+  delete this.spaces[parentId]
+  parentSpace.remove()
+
+  // Enable hooking to removals
+  this.emit('close', { id: parentId })
 }
