+++ lib/loaders/TreeLoader/index.js
@@ -1,36 +1,36 @@
 const emitter = require('component-emitter')
 
-const TreeLoader = function (config) {
-  // @tapspace.loaders.TreeLoader(config)
-  // @TreeLoader
+const EventTreeLoader = function (config) {
+  // @tapspace.loaders.EventTreeLoader(config)
+  // @EventTreeLoader
   //
   // An asynchronous and recursive loader for your tree-structured content.
   // Use this to build infinite or very deep, zoomable web applications
   // to overcome limits of floating point arithmetics.
   //
-  // To ignite the loader, you need to implement a few functions.
+  // To setup the loader, you need to implement a few functions:
+  // mapper, backmapper, tracker, and backtracker.
+  // The mappers define the relative positions of content and
+  // the trackers define the order of content.
+  // The mappers are optional in a sense that you need to implement only
+  // one of the two for the loader to work.
+  // Pick the one that best suits your data: if your datum stores
+  // its location relative to the parent, implement a backmapper;
+  // if your datum stores the locations of its children, implement a mapper.
   //
   // Parameters:
   //   config
   //     an object with properties:
   //       viewport
   //         a tapspace.components.Viewport
-  //       placeholder
-  //         a function (id), synchronous. Render a placeholder element
-  //         .. that is displayed during fetching of the actual content.
-  //         .. Return a Component.
-  //       fetcher
-  //         a function (id, callback), asynchronous.
-  //         .. Fetch the content for id from your data store,
-  //         .. store it locally somehow, and then call the callback.
-  //         .. The callback is a function (err).
-  //       generator
-  //         a function (id), synchronous. Render the content element.
-  //         .. Called after a successful fetch. Return a Component.
   //       mapper
-  //         a function (parentId, parent, childId), synchronous, where
+  //         optional function (parentId, parent, childId), synchronous, where
   //         .. parent is a Component. Determine the placement of the child
   //         .. by returning a Basis relative to the parent.
+  //       backmapper
+  //         optional function (childId, child, parentId), synchronous, where
+  //         .. child is a Component. Determine the placement of the parent
+  //         .. by returning a Basis relative to the child.
   //       tracker
   //         a function (parentId, parent), synchronous.
   //         .. The parent is a Component.
@@ -42,113 +42,151 @@ const TreeLoader = function (config) {
   //         .. Return the parent ID string.
   //         .. Return null if the child does not have a parent.
   //
-  // Here is a minimal function for it:
+  // By calling `loader.init(id, basis)` the loader is initiated and
+  // the tree-building starts.
+  // However, the loader does not yet know your content.
+  // To provide the content, and also means to render and retrieve the content,
+  // you need to set an 'open' event handler:
   //
   // ```
-  // const store = {}
+  // loader.on('open', (id) => {
+  //   ...
+  // })
+  // ```
+  //
+  // Inside the handler, you can create your content for your tree node,
+  // like text, images, html, loading animations.
+  // Place the content into a tapspace item or plane,
+  // and call `loader.open(id, content)`.
+  // You can also call `loader.placeholder(id, content)`
+  // to place a placeholder that will be replaced when the actual content
+  // is retrieved and opened.
+  //
+  // Here is a minimal example:
+  //
+  // ```
+  // loader.on('open', (id) => {
+  //   loader.placeholder(id, '...')
+  //   fetch('https://api.example.com/data')
+  //     .then(response => {
+  //       if (!response.ok) throw new Error('Network response was not OK')
+  //       return response.json()
+  //     })
+  //     .then(data => {
+  //       loader.open(id, data.html)
+  //       callback(null)
+  //     })
+  //     .catch(error => {
+  //       loader.open(id, error.message)
+  //       console.error('Error:', error)
+  //       callback(error)
+  //     })
+  // })
+  // ```
+  //
+  // The loader also emits 'close' event.
+  // While the loader closes content automatically,
+  // if your code needs some deconstruction behavior,
+  // place a listener. For example:
+  //
+  // ```
+  // loader.on('close', (id) => {
+  //   streams[id].close()
+  // })
+  // ```
+  //
+  // Here is a complete configuration example:
+  //
+  // ```
   // const loader = new tapspace.loaders.TreeLoader({
   //   viewport: viewport,
   //
-  //   placeholder: function (id) {
-  //     // Render placeholder
-  //     const plane = tapspace.createPlane()
-  //     return plane
-  //   },
-  //
-  //   fetcher: function (id, callback) {
-  //     // Call your API or data store here.
-  //     fetch('https://api.example.com/data')
-  //       .then(response => {
-  //         if (!response.ok) {
-  //           throw new Error('Network response was not OK')
-  //         }
-  //         return response.json()
-  //       })
-  //       .then(data => {
-  //         console.log('Data:', data)
-  //         store[id] = data
-  //         callback(null)
-  //       })
-  //       .catch(error => {
-  //         console.error('Error:', error)
-  //         callback(error)
-  //       })
-  //   },
-  //
-  //   generator: function (id) {
-  //     // Render content to replace the placeholder
-  //     const label = store[id].title
-  //     const plane = tapspace.createPlane()
-  //     plane.html(label)
-  //     return plane
-  //   },
-  //
   //   mapper: function (parentId, parent, childId) {
-  //     // Place the child relative to the parent.
-  //     // In other words, find a basis for the child.
+  //     // Find the location for the child, relative to the parent.
+  //     // The return value must be a Basis.
+  //     // Get the parent basis, and transform it as you like.
   //     // Return null if there is no place for the child.
+  //     const data = store[parentId]
+  //     const dataPoint = data.points[childId]
+  //     if (dataPoint) {
+  //       return parent.getBasis().offset(dataPoint.x, dataPoint.y)
+  //     }
+  //     return null
+  //   },
   //
-  //     return basis
+  //   backmapper: function (childId, child, parentId) {
+  //     // Find the location for the parent, relative to the child.
+  //     // The mapper and backmapper are optional, but you need to implement
+  //     // at least one of the two.
+  //     const data = store[childId]
+  //     const dataPoint = data.parent.point
+  //     if (dataPoint) {
+  //       return child.getBasis().offset(dataPoint.x, dataPoint.y)
+  //     }
   //     return null
   //   },
   //
   //   tracker: function (parentId, parent) {
-  //     // Get IDs of the children of the parent component.
-  //     // Return empty array if there are no childs.
-  //     const data = store[parentId]
-  //     const childrenIds = data.points.map(p => p.id)
-  //     return childrenIds
+  //     // Find IDs of the child nodes, given the parent node.
+  //     // Return an empty array if there are no childred.
+  //     const dataPoints = store[parentId].points
+  //     const idArray = dataPoints.map(p => p.id)
+  //     return idArray
   //   },
   //
-  //   backtracker: function (childId) {
-  //     // Find parent id.
-  //     // Return null if there is no parent i.e. child is the root.
+  //   backtracker: function (childId, child) {
+  //     // Find ID of the parent node, given the child node.
+  //     // Return null if there are no parents. Such child is the root node.
   //     const data = store[childId]
-  //     if (data.parentId) {
-  //       return data.parentId
+  //     if (data.parent.id) {
+  //       return data.parent.id
   //     }
   //     return null
   //   }
   // })
   // ```
   //
+  // Now you have the TreeLoader constructed.
+  // You still need to build a driver for it. See examples.
+  //
   this.viewport = config.viewport
-  this.placeholder = config.placeholder
-  this.fetcher = config.fetcher
-  this.generator = config.generator
+
   this.mapper = config.mapper
+  this.backmapper = config.backmapper || null
   this.tracker = config.tracker
   this.backtracker = config.backtracker
 
-  // Track which spaces are loading.
-  // id -> Component
-  this.placeholders = {}
-
-  // Track which spaces are finished loading.
+  // Track created spaces.
   // id -> Component
   this.spaces = {}
 
-  // Callbacks that wait for spaces to finish.
-  // id -> array of function
-  this.callbacks = {}
+  // Cached bases, at least for the init basis.
+  this.bases = {}
+
+  // Track which spaces are still loading.
+  this.loading = {}
+
+  // Track to which depth the node should open.
+  this.demand = {}
 }
 
-module.exports = TreeLoader
-const proto = TreeLoader.prototype
-proto.isTreeLoader = true
+module.exports = EventTreeLoader
+const proto = EventTreeLoader.prototype
+proto.isEventTreeLoader = true
 
 // Inherit
 emitter(proto)
 
 proto.init = require('./init')
+proto.closeChild = require('./closeChild')
 proto.closeChildren = require('./closeChildren')
 proto.closeNeighbors = require('./closeNeighbors')
 proto.closeParent = require('./closeParent')
+proto.open = require('./open')
 proto.openChild = require('./openChild')
 proto.openChildren = require('./openChildren')
 proto.openNeighbors = require('./openNeighbors')
 proto.openParent = require('./openParent')
-proto.registerCallback = require('./registerCallback')
+proto.placeholder = require('./placeholder')
 proto.remapChildren = require('./remapChildren')
-proto.resolveCallbacks = require('./resolveCallbacks')
 proto.removeSpace = require('./removeSpace')
