+++ lib/loaders/TreeLoader/init.js
@@ -1,5 +1,5 @@
-module.exports = function (id, basis, callback) {
-  // @TreeLoader:init(id, basis, callback)
+module.exports = function (id, basis) {
+  // @TreeLoader:init(id, basis)
   //
   // Initialize the tree. Add and load the first space.
   //
@@ -8,38 +8,20 @@ module.exports = function (id, basis, callback) {
   //     a string
   //   basis
   //     a Basis, the placement of the first item.
-  //   callback
-  //     a function (err, spaceId, space)
+  //
+  // Return
+  //   this, for chaining
   //
 
-  if (callback) {
-    this.registerCallback(id, callback)
-  }
+  // Register basis
+  this.bases[id] = basis
+  // Register demand
+  this.demand[id] = 1
 
-  const placeholder = this.placeholder(id)
-  this.placeholders[id] = placeholder
-  this.viewport.addChild(placeholder)
-  placeholder.setBasis(basis)
+  setTimeout(() => {
+    this.loading[id] = true
+    this.emit('open', id)
+  }, 0)
 
-  this.fetcher(id, (err) => {
-    if (err) {
-      return this.resolveCallbacks(id, err)
-    }
-
-    const postSpace = this.spaces[id]
-    if (!postSpace) {
-      const space = this.generator(id)
-      this.spaces[id] = space
-      this.viewport.addChild(space)
-      space.setBasis(basis)
-    }
-
-    const postPlaceholder = this.placeholders[id]
-    if (postPlaceholder) {
-      delete this.placeholders[id]
-      postPlaceholder.remove()
-    }
-
-    return this.resolveCallbacks(id)
-  })
+  return this
 }
