+++ lib/loaders/TreeLoader/open.js
