+++ lib/loaders/TreeLoader/openNeighbors.js
@@ -1,8 +1,8 @@
 const findBacktier = require('./utils/findBacktier')
 const findFrontier = require('./utils/findFrontier')
 
-module.exports = function (id, maxDepth, callback) {
-  // @TreeLoader:openNeighbors(id, maxDepth, callback)
+module.exports = function (id, maxDepth) {
+  // @TreeLoader:openNeighbors(id, maxDepth)
   //
   // Open parents and children until max depth.
   //
@@ -11,51 +11,25 @@ module.exports = function (id, maxDepth, callback) {
   //     a string, the first node ID.
   //   maxDepth
   //     a number, the maximum depth. For example `2`
-  //   callback
-  //     a function (err), called once all opened or at error.
   //
-  if (!callback) {
-    callback = () => {}
-  }
 
   // Stop recursion.
   if (maxDepth <= 0) {
-    setTimeout(() => {
-      callback(null)
-    }, 0)
-
     return
   }
 
   // Space may not exist, yet.
   const space = this.spaces[id]
   if (!space) {
-    setTimeout(() => {
-      callback(null)
-    }, 0)
-
     return
   }
 
-  let errored = false
-  let numCallbacks = 0
-  let expectedNumCallbacks = 0
-  const requestCallback = (err) => {
-    if (errored || numCallbacks >= expectedNumCallbacks) {
-      return
-    }
-
-    if (err) {
-      errored = true
-      return callback(err)
-    }
-
-    numCallbacks += 1
-
-    if (numCallbacks === expectedNumCallbacks) {
-      callback(null)
-    }
+  // Update demand for this node
+  let localDemand = maxDepth + 1
+  if (this.demand[id]) {
+    localDemand = Math.max(localDemand, this.demand[id])
   }
+  this.demand[id] = localDemand
 
   // Find neighbors that have unopened children or parents.
   const frontier = findFrontier(
@@ -72,46 +46,21 @@ module.exports = function (id, maxDepth, callback) {
     maxDepth - 1
   )
 
-  expectedNumCallbacks += frontier.length
-
   frontier.forEach((finding) => {
-    this.openChildren(finding.id, (err, openedIds) => {
-      if (err) {
-        return requestCallback(err)
-      }
-
-      if (finding.depth + 1 < maxDepth) {
-        openedIds.forEach((openedId) => {
-          if (openedId) {
-            expectedNumCallbacks += 1
-            const subDepth = maxDepth - finding.depth - 1
-            // Recursion
-            this.openNeighbors(openedId, subDepth, requestCallback)
-          }
-        })
-      }
-
-      return requestCallback(null)
-    })
+    // Increase neighborhood demand.
+    const currentDemand = this.demand[finding.id] || 0
+    const minDemand = localDemand - finding.depth
+    this.demand[finding.id] = Math.max(minDemand, currentDemand)
+    // Open children
+    this.openChildren(finding.id)
   })
 
-  expectedNumCallbacks += backtier.length
-
   backtier.forEach((finding) => {
-    this.openParent(finding.id, (err, parentId) => {
-      if (err) {
-        return requestCallback(err)
-      }
-
-      if (parentId && finding.depth + 1 < maxDepth) {
-        // Recursion
-        expectedNumCallbacks += 1
-
-        const subDepth = maxDepth - finding.depth - 1
-        this.openNeighbors(parentId, subDepth, requestCallback)
-      }
-
-      return requestCallback(null)
-    })
+    // Increase neighborhood demand.
+    const currentDemand = this.demand[finding.id] || 0
+    const minDemand = localDemand - finding.depth
+    this.demand[finding.id] = Math.max(minDemand, currentDemand)
+    // Open parent
+    this.openParent(finding.id)
   })
 }
