+++ lib/loaders/TreeLoader/openParent.js
@@ -1,22 +1,16 @@
-module.exports = function (childId, callback) {
-  // @TreeLoader:openParent(childId, callback)
+module.exports = function (childId) {
+  // @TreeLoader:openParent(childId)
   //
   // Open a parent space relative to the given child.
   //
   // Parameters:
   //   childId
   //     a string
-  //   callback
-  //     a function (err, parentId, parent)
   //
 
   // Child may not exist anymore.
   const childSpace = this.spaces[childId]
   if (!childSpace) {
-    setTimeout(() => {
-      callback(null, null, null)
-    }, 0)
-
     return
   }
 
@@ -24,70 +18,30 @@ module.exports = function (childId, callback) {
 
   // Root has no parent.
   if (!parentId) {
-    setTimeout(() => {
-      callback(null, null, null)
-    }, 0)
-
-    return
-  }
-
-  if (callback) {
-    this.registerCallback(parentId, callback)
-  }
-
-  // Placeholder for parent may exist already.
-  if (this.placeholders[parentId]) {
-    // Rely that the callback will be resolved when the parent is finished.
     return
   }
 
   // Parent may exist already.
   const parentSpace = this.spaces[parentId]
   if (parentSpace) {
-    return this.resolveCallbacks(parentId)
+    return
   }
 
-  // Construct
-  const placeholder = this.placeholder(parentId)
-  this.placeholders[parentId] = placeholder
-  this.viewport.addChild(placeholder)
-  // Place
-  const childBasis = childSpace.getBasis()
-  placeholder.setBasis(childBasis)
+  // The parent may be already loading.
+  if (this.loading[parentId]) {
+    return
+  }
 
-  this.fetcher(parentId, (err) => {
-    // In any case, clean up placeholder
-    const postPlaceholder = this.placeholders[parentId]
-    if (postPlaceholder) {
-      delete this.placeholders[parentId]
-      postPlaceholder.remove()
-    }
+  // The child may have low demand.
+  const childDemand = this.demand[childId] || 0
+  if (childDemand < 2) {
+    return
+  }
 
-    if (err) {
-      return this.resolveCallbacks(parentId, err)
-    }
+  // Set demand for the non-existing parent
+  this.demand[parentId] = childDemand - 1
+  // Begin loading
+  this.loading[parentId] = true
 
-    // Child may not exist anymore.
-    const postChildSpace = this.spaces[childId]
-    if (!postChildSpace) {
-      return this.resolveCallbacks(parentId)
-    }
-    // Child may have been replaced. Find the basis again.
-    const postChildBasis = postChildSpace.getBasis()
-
-    // Construct
-    const parentSpace = this.generator(parentId)
-    // Find position
-    const insertionBasis = this.mapper(parentId, parentSpace, childId)
-    if (!insertionBasis) {
-      // No basis for the child in the map, currently.
-      return this.resolveCallbacks(parentId)
-    }
-    // Place
-    this.spaces[parentId] = parentSpace
-    this.viewport.addChild(parentSpace)
-    parentSpace.matchBasis(insertionBasis, postChildBasis)
-
-    return this.resolveCallbacks(parentId)
-  })
+  this.emit('open', parentId)
 }
