+++ lib/loaders/TreeLoader/placeholder.js
