--- lib/loaders/TreeLoader/registerCallback.js
@@ -1,13 +1,0 @@
-module.exports = function (id, cb) {
-  // FOR INTERNAL USE
-  //
-  // Register a callback.
-  // Modifies the callbacks object.
-  //
-  const cbs = this.callbacks[id]
-  if (cbs) {
-    cbs.push(cb)
-  } else {
-    this.callbacks[id] = [cb]
-  }
-}
