+++ lib/loaders/TreeLoader/removeSpace.js
@@ -9,15 +9,12 @@ module.exports = function (spaceId) {
   //     a string
   //
 
-  const placeholder = this.placeholders[spaceId]
-  if (placeholder) {
-    delete this.placeholders[spaceId]
-    placeholder.remove()
-  }
-
   const space = this.spaces[spaceId]
   if (space) {
     delete this.spaces[spaceId]
+    delete this.demand[spaceId]
     space.remove()
+    // Enable hooking to removals
+    this.emit('close', { id: spaceId })
   }
 }
