+++ lib/loaders/TreeLoader/utils/findFrontier.js
@@ -55,7 +55,7 @@ module.exports = (spaces, tracker, id, maxDepth) => {
           depths.push(depth + 1)
         }
       } else if (!cursorInFrontier) {
-        // Not every children is open: cursor belongs to the frontier.
+        // Not every child is open: cursor belongs to the frontier.
         frontier.push({
           id: cursor,
           space: cursorSpace,
