+++ lib/loaders/TreeLoader/utils/findTreeDistances.js
