+++ lib/loaders/index.js
@@ -5,7 +5,6 @@
 // loading, construction, and destruction.
 //
 
-exports.EventTreeLoader = require('./EventTreeLoader')
 exports.FractalLoader = require('./FractalLoader')
 exports.TreeLoader = require('./TreeLoader')
 exports.loadImages = require('./loadImages')
